separate PluginTask for readability
